Move KeyMId into outer class.
